Replace start scripts with a single Python program
This change replaces the set of shell scripts and Python scripts to
launch Kolibri with a single integrated Python application that manages
the Kolibri service and desktop launcher. This script runs Kolbiri,
a new idle killer service, and a redirect / loading screen service, in
separate threads.

The idle killer service is always run alongside the Kolbri service, on a
randomly assigned port number which is passed to Kolibri.

Each time the application is launched on the desktop, a separate
redirect service is created specifically for that instance, again on a
randomly assigned port number.

With this change, the flatpak can be used in an environment where the
Kolbri service is always available on the system via socket activation.