Replace flatpak-specific code with kolibri-installer-gnome
This removes most of the programs and patches that were bundled with
this package metadata. Instead, we use the kolibri-installer-gnome
project, which provides most of the mechanisms we require. There is a
single script maintained here, _run-kolibri-gnome.sh_. This script is
needed to move Kolibri's data directory to $XDG_DATA_HOME.