Switch to yaml and comment out branch for git sources
The `branch` property was being misused here. When it is used along with
the `commit` property, flatpak-builder verifies that the branch is
pointing at the specified commit and fails if it is not. Instead, we
want to point at a specific commit, noting the branch for documentation
reasons. This is easier to do with YAML because it supports comments.