Move JavaScript source maps to debug extension
These are not needed in normal use of Kolibri, and there are 30MB of
them: more than 10% of the size of the entire
org.learningequality.Kolibri Flatpak!

Move them to /app/lib/debug, and add symlinks in their place.
/app/lib/debug is automatically split into the
org.learningequality.Kolibri.Debug extension, which is not installed by
default.